change of the notification ID data type from NSInteger to uint32

+++ bindings/ios/MEGASdk.mm
@@ -1208,7 +1208,7 @@ using namespace mega;
     }
 }
 
-- (void)setLastReadNotificationWithNotificationId:(NSInteger)notificationId delegate:(id<MEGARequestDelegate>)delegate {
+- (void)setLastReadNotificationWithNotificationId:(uint32_t)notificationId delegate:(id<MEGARequestDelegate>)delegate {
     if (self.megaApi) {
         self.megaApi->setLastReadNotification((int)notificationId, [self createDelegateMEGARequestListener:delegate singleListener:YES]);
     }
